app/flask_app/requirements.txt
Flask==2.2.5              # Flask web framework
Flask-SQLAlchemy==3.0.5   # For database integration
Flask-Migrate==4.0.4      # For database migrations (optional)
Flask-Cors==3.0.10        # To handle cross-origin requests (for React frontend)
PyJWT==2.8.0              # For JSON Web Token (JWT) authentication
python-dotenv==1.0.0      # To load environment variables
